Move json_schema from fsm to types
