+++ outlines/types/__init__.py
@@ -5,7 +5,6 @@ from typing import Union
 
 from jsonschema import Draft202012Validator as Validator
 from jsonschema.exceptions import SchemaError
-from outlines_core.fsm.json_schema import build_regex_from_schema
 from pydantic import BaseModel, TypeAdapter
 from typing_extensions import _TypedDictMeta  # type: ignore
